requirements.txt
streamlit==1.28.0
deepface==0.0.79
opencv-python-headless==4.8.1.78
tensorflow>=2.16.0
numpy==1.24.3
Pillow==10.0.1
matplotlib==3.7.3
protobuf==3.20.3